safari-600.7-branch/LayoutTests/http/tests/notifications/legacy/window-show-on-click-expected.txt

This test checks that a script can open windows when the user clicks on a native notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS webkitNotifications.checkPermission() is 0
PASS clicking on notification successfully opened a window
PASS successfullyParsed is true

TEST COMPLETE
